Move date part rendering tests to separate module
This moves the date part rendering tests out of the
test_dataflow_to_sql_plan module into their own space.

The new module is tagged for both date_part and granularity
query rendering, because impending updates to date_part and
granularity handling will likely require some additional
query rendering tests, and date_part and granularity rendering
will be fairly similar in terms of the pattern of the final output.
